more work on pathfinding (not in sim yet)
total: 15722
correct: 64.0%
time_outs: 0.5%
improvement: 27.9%
log loss: 0.276955

+++ simulator/src/sim/actions/attack.rs
@@ -47,11 +47,11 @@ impl AbilityImpl for AttackImpl {
     }
 }
 
-pub fn attack_range(user: &Combatant) -> i8 {
+pub fn attack_range(user: &Combatant) -> u8 {
     if user.frog() || user.berserk() && user.monster() {
         1
     } else {
-        user.main_hand().map_or(1, |eq| eq.range)
+        user.main_hand().map_or(1, |eq| eq.range as u8)
     }
 }
